splitting fixed

+++ core/src/main/java/org/apache/mahout/cf/taste/impl/eval/AbstractDifferenceRecommenderEvaluator.java
@@ -112,7 +112,7 @@ public abstract class AbstractDifferenceRecommenderEvaluator implements Recommen
 
 		LongPrimitiveIterator it = dataModel.getUserIDs();
 		while (it.hasNext()) {
-			long userID = it.nextLong();
+			long userID = it.nextInt();
 			if (random.nextDouble() < evaluationPercentage) {
 				splitOneUsersPrefs(trainingPercentage, trainingPrefs, testPrefs, userID, dataModel);
 			}
