+++ core/src/main/java/org/apache/mahout/cf/taste/impl/recommender/svd/RatingSGDFactorizer.java
@@ -107,7 +107,7 @@ public class RatingSGDFactorizer extends AbstractFactorizer {
 		int numPreferences = 0;
 		LongPrimitiveIterator userIDs = dataModel.getUserIDs();
 		while (userIDs.hasNext()) {
-			PreferenceArray preferencesFromUser = dataModel.getPreferencesFromUser(userIDs.nextLong());
+			PreferenceArray preferencesFromUser = dataModel.getPreferencesFromUser(userIDs.nextInt());
 			numPreferences += preferencesFromUser.length();
 		}
 		return numPreferences;
@@ -121,7 +121,7 @@ public class RatingSGDFactorizer extends AbstractFactorizer {
 		LongPrimitiveIterator userIDs = dataModel.getUserIDs();
 		int index = 0;
 		while (userIDs.hasNext()) {
-			long userID = userIDs.nextLong();
+			long userID = userIDs.nextInt();
 			PreferenceArray preferencesFromUser = dataModel.getPreferencesFromUser(userID);
 			for (Preference preference : preferencesFromUser) {
 				cachedUserIDs[index] = userID;
@@ -172,7 +172,7 @@ public class RatingSGDFactorizer extends AbstractFactorizer {
 		RunningAverage average = new FullRunningAverage();
 		LongPrimitiveIterator it = dataModel.getUserIDs();
 		while (it.hasNext()) {
-			for (Preference pref : dataModel.getPreferencesFromUser(it.nextLong())) {
+			for (Preference pref : dataModel.getPreferencesFromUser(it.nextInt())) {
 				average.addDatum(pref.getValue());
 			}
 		}
