+++ examples/src/main/java/org/apache/mahout/cf/taste/example/kddcup/track2/TrackItemSimilarity.java
@@ -90,7 +90,7 @@ final class TrackItemSimilarity implements ItemSimilarity {
 		FastIDSet allSimilarItemIDs = new FastIDSet();
 		LongPrimitiveIterator allItemIDs = trackData.keySetIterator();
 		while (allItemIDs.hasNext()) {
-			long possiblySimilarItemID = allItemIDs.nextLong();
+			long possiblySimilarItemID = allItemIDs.nextInt();
 			if (!Double.isNaN(itemSimilarity(itemID, possiblySimilarItemID))) {
 				allSimilarItemIDs.add(possiblySimilarItemID);
 			}
